+++ juntagrico/admins/filters.py
@@ -1,7 +1,7 @@
 import datetime
 
-from django.utils import timezone
 from django.contrib.admin.filters import DateFieldListFilter
+from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
