+++ juntagrico/admins/job_admin.py
@@ -4,12 +4,12 @@ from django.http import HttpResponseRedirect
 from django.utils.translation import gettext as _
 
 from juntagrico.admins import BaseAdmin
+from juntagrico.admins.filters import FutureDateTimeFilter
 from juntagrico.admins.forms.job_copy_form import JobCopyForm
 from juntagrico.admins.inlines.assignment_inline import AssignmentInline
 from juntagrico.dao.jobtypedao import JobTypeDao
 from juntagrico.entity.jobs import RecuringJob
 from juntagrico.util.admin import formfield_for_coordinator, queryset_for_coordinator, extra_context_for_past_jobs
-from juntagrico.admins.filters import FutureDateTimeFilter
 
 
 class JobAdmin(BaseAdmin):
