+++ juntagrico/config.py
@@ -1,6 +1,5 @@
-from django.utils.translation import gettext as _
-
 from django.conf import settings
+from django.utils.translation import gettext as _
 
 
 def _get_setting(setting_key, default: object = ''):
