+++ juntagrico/lifecycle/job.py
@@ -2,7 +2,7 @@ from django.core.exceptions import ValidationError
 from django.utils.translation import gettext as _
 
 from juntagrico.dao.assignmentdao import AssignmentDao
-from juntagrico.mailer import MemberNotification
+from juntagrico.mailer import membernotification
 from juntagrico.signals import job_canceled, job_time_changed
 
 
@@ -22,14 +22,14 @@ def handle_job_canceled(sender, instance, **kwargs):
         assignment.delete()
     instance.slots = 0
     if len(emails) > 0:
-        MemberNotification.job_canceled(emails, instance)
+        membernotification.job_canceled(emails, instance)
 
 
 def handle_job_time_changed(sender, instance, **kwargs):
     assignments = AssignmentDao.assignments_for_job(instance.id)
     emails = set([assignment.member.email for assignment in assignments])
     if len(emails) > 0:
-        MemberNotification.job_time_changed(emails, instance)
+        membernotification.job_time_changed(emails, instance)
 
 
 def check_job_consistency(instance):
