+++ juntagrico/management/commands/generate_depot_list.py
@@ -7,9 +7,8 @@ from juntagrico.dao.extrasubscriptioncategorydao import ExtraSubscriptionCategor
 from juntagrico.dao.listmessagedao import ListMessageDao
 from juntagrico.dao.subscriptiondao import SubscriptionDao
 from juntagrico.dao.subscriptionproductdao import SubscriptionProductDao
-from juntagrico.mailer import MemberNotification
+from juntagrico.mailer import membernotification
 from juntagrico.util.pdf import render_to_pdf_storage
-
 from juntagrico.util.temporal import weekdays
 
 
@@ -48,7 +47,7 @@ class Command(BaseCommand):
                 emails = []
                 for member in subscription.recipients:
                     emails.append(member.email)
-                MemberNotification.depot_changed(emails, subscription.depot)
+                membernotification.depot_changed(emails, subscription.depot)
 
         if options['force'] and not options['future']:
             print('future depots ignored, use --future to override')
