+++ juntagrico/management/commands/generate_testdata_advanced.py
@@ -1,5 +1,4 @@
 ﻿import json
-import math
 import ssl
 import sys
 import time as mytime
@@ -7,6 +6,7 @@ import urllib.error
 import urllib.parse
 import urllib.request
 
+import math
 from django.core.management.base import BaseCommand
 from django.utils import timezone
 from django.utils.text import slugify
