+++ juntagrico/management/commands/remind_members.py
@@ -3,9 +3,9 @@ import datetime
 from django.core.management.base import BaseCommand
 from django.utils import timezone
 
-from juntagrico.mailer import MemberNotification
 from juntagrico.dao.assignmentdao import AssignmentDao
 from juntagrico.dao.jobdao import JobDao
+from juntagrico.mailer import membernotification
 
 
 class Command(BaseCommand):
@@ -21,7 +21,7 @@ class Command(BaseCommand):
                 if assignment.member is not None:
                     participants.append(str(assignment.member))
                     emails.append(assignment.member.email)
-            MemberNotification.job_reminder(emails, job, ', '.join(participants))
+            membernotification.job_reminder(emails, job, ', '.join(participants))
             job.reminder_sent = True
             job.save()
             print(('reminder sent for job ' + str(job.id)))
