+++ juntagrico/util/decorators.py
@@ -1,0 +1,7 @@
+
+
+def chainable(func):
+    def wrapper(*args, **kwargs):
+        func(*args, **kwargs)
+        return args[0]
+    return wrapper
