+++ juntagrico/util/temporal.py
@@ -3,7 +3,6 @@ import datetime
 from datetime import timedelta
 
 from django.utils import timezone
-from django.utils.timezone import get_default_timezone as gdtz
 from django.utils.translation import gettext as _
 
 from juntagrico.config import Config
